* New feature: Support for ESP8266. * Fixed issue #5: error: invalid conversion from 'long unsigned int (*)()' to 'SoftTimer::CounterFunctionPointer {aka unsigned int (*)()}' [-fpermissive]

+++ src/SoftTimers.h
@@ -17,7 +17,7 @@ public:
    * Description:
    *   Defines a generic 32 bits counting function.
    ****************************************************************************/
-  typedef uint32_t (*CounterFunctionPointer)();
+  typedef unsigned long (*CounterFunctionPointer)();
 
   /****************************************************************************
    * Description:
